requirements.txt
streamlit>=1.35.0
pandas>=2.0.0
numpy>=1.24.0
pdfplumber>=0.10.0
PyPDF2>=3.0.0
openpyxl>=3.1.0
xlrd>=2.0.1
docx2txt>=0.8
scikit-learn>=1.3.0
openai>=1.0.0
google-generativeai>=0.5.0
anthropic>=0.25.0
python-dotenv>=1.0.0
